requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.3.0
xgboost>=1.7.0
scipy>=1.9.0
click>=8.0.0
pyyaml>=6.0
joblib>=1.2.0
pytest>=7.0.0
ViennaRNA>=2.5.0
forgi>=2.0.0
torch>=2.0.0
transformers>=4.21.0
tokenizers>=0.13.0
huggingface-hub>=0.15.0